+++ src/Sentry/ITransactionContext.cs
@@ -1,0 +1,13 @@
+﻿namespace Sentry
+{
+    /// <summary>
+    /// Transaction metadata.
+    /// </summary>
+    public interface ITransactionContext : ISpanContext
+    {
+        /// <summary>
+        /// Transaction name.
+        /// </summary>
+        string Name { get; }
+    }
+}