+++ test/Sentry.Tests/Internals/MainSentryEventProcessorTests.cs
@@ -216,9 +216,9 @@ namespace Sentry.Tests.Internals
         }
 
         [Theory]
-        [InlineData(null, Constants.ProductionEnvironmentSetting)] // Missing: will get default value.
-        [InlineData("", Constants.ProductionEnvironmentSetting)] // Missing: will get default value.
-        [InlineData(" ", Constants.ProductionEnvironmentSetting)] // Missing: will get default value.
+        [InlineData(null, Sentry.Internal.Constants.ProductionEnvironmentSetting)] // Missing: will get default value.
+        [InlineData("", Sentry.Internal.Constants.ProductionEnvironmentSetting)] // Missing: will get default value.
+        [InlineData(" ", Sentry.Internal.Constants.ProductionEnvironmentSetting)] // Missing: will get default value.
         [InlineData("a", "a")] // Provided: nothing will change.
         [InlineData("production", "production")] // Provided: nothing will change. (value has a lower case 'p', different to default value)
         [InlineData("aBcDe F !@#$ gHi", "aBcDe F !@#$ gHi")] // Provided: nothing will change. (Case check)
@@ -234,9 +234,9 @@ namespace Sentry.Tests.Internals
         }
 
         [Theory]
-        [InlineData(null, Constants.ProductionEnvironmentSetting)] // Missing: will get default value.
-        [InlineData("", Constants.ProductionEnvironmentSetting)] // Missing: will get default value.
-        [InlineData(" ", Constants.ProductionEnvironmentSetting)] // Missing: will get default value.
+        [InlineData(null, Sentry.Internal.Constants.ProductionEnvironmentSetting)] // Missing: will get default value.
+        [InlineData("", Sentry.Internal.Constants.ProductionEnvironmentSetting)] // Missing: will get default value.
+        [InlineData(" ", Sentry.Internal.Constants.ProductionEnvironmentSetting)] // Missing: will get default value.
         [InlineData("a", "a")] // Provided: nothing will change.
         [InlineData("Production", "Production")] // Provided: nothing will change. (value has a upper case 'p', different to default value)
         [InlineData("aBcDe F !@#$ gHi", "aBcDe F !@#$ gHi")] // Provided: nothing will change. (Case check)
@@ -246,7 +246,7 @@ namespace Sentry.Tests.Internals
             var evt = new SentryEvent();
 
             EnvironmentVariableGuard.WithVariable(
-                Constants.EnvironmentEnvironmentVariable,
+                Sentry.Internal.Constants.EnvironmentEnvironmentVariable,
                 environment,
                 () =>
                 {
@@ -295,7 +295,7 @@ namespace Sentry.Tests.Internals
             var evt = new SentryEvent();
             _ = sut.Process(evt);
 
-            Assert.Equal(Sentry.Protocol.Constants.Platform, evt.Platform);
+            Assert.Equal(Constants.Platform, evt.Platform);
         }
 
         [Fact]
@@ -327,7 +327,7 @@ namespace Sentry.Tests.Internals
 
             _ = sut.Process(evt);
 
-            Assert.Equal(Constants.SdkName, evt.Sdk.Name);
+            Assert.Equal(Sentry.Internal.Constants.SdkName, evt.Sdk.Name);
             Assert.Equal(typeof(ISentryClient).Assembly.GetNameAndVersion().Version, evt.Sdk.Version);
         }